refactor(api): extract browser user contracts
