+++ docs/architecture/packages.md
@@ -43,6 +43,7 @@ leaf-safe surface. They currently own the dependency-free-or-`zod` subset that c
 `@promptfoo/schema` package:
 
 - shared token/input contracts
+- browser-safe common and user API DTOs
 - provider environment override schema
 - prompt contracts and prompt validation
 - transform contracts and shared transform validation
