+++ test/contracts/shimEquivalence.test.ts
@@ -1,10 +1,14 @@
 import { describe, expect, it } from 'vitest';
+import * as contractsApiCommon from '../../src/contracts/api/common';
+import * as contractsApiUser from '../../src/contracts/api/user';
 import * as contractsEnv from '../../src/contracts/env';
 import * as contractsPrompts from '../../src/contracts/prompts';
 import * as contractsShared from '../../src/contracts/shared';
 import * as contractsTransform from '../../src/contracts/transform';
 import * as contractsValidatorPrompts from '../../src/contracts/validators/prompts';
 import * as contractsValidatorShared from '../../src/contracts/validators/shared';
+import * as legacyApiCommon from '../../src/types/api/common';
+import * as legacyApiUser from '../../src/types/api/user';
 import * as legacyEnv from '../../src/types/env';
 import * as legacyPrompts from '../../src/types/prompts';
 import * as legacyShared from '../../src/types/shared';
@@ -19,6 +23,8 @@ import * as legacyValidatorShared from '../../src/validators/shared';
  */
 describe('legacy shim equivalence', () => {
   const pairs: Array<[string, Record<string, unknown>, Record<string, unknown>]> = [
+    ['api/common', legacyApiCommon, contractsApiCommon],
+    ['api/user', legacyApiUser, contractsApiUser],
     ['env', legacyEnv, contractsEnv],
     ['prompts', legacyPrompts, contractsPrompts],
     ['shared', legacyShared, contractsShared],
